+++ Source/Scene/BillboardCollection.js
@@ -60,6 +60,7 @@ define([
     var COLOR_INDEX = Billboard.COLOR_INDEX;
     var ROTATION_INDEX = Billboard.ROTATION_INDEX;
     var ALIGNED_AXIS_INDEX = Billboard.ALIGNED_AXIS_INDEX;
+    var SCALE_BY_DISTANCE_INDEX = Billboard.SCALE_BY_DISTANCE_INDEX;
     var NUMBER_OF_PROPERTIES = Billboard.NUMBER_OF_PROPERTIES;
 
     // PERFORMANCE_IDEA:  Use vertex compression so we don't run out of
@@ -74,7 +75,8 @@ define([
         direction : 6,
         pickColor : 7,  // pickColor and color shared an index because pickColor is only used during
         color : 7,      // the 'pick' pass and 'color' is only used during the 'color' pass.
-        rotationAndAlignedAxis : 8
+        rotationAndAlignedAxis : 8,
+        scaleByDistance : 9
     };
 
     // Identifies to the VertexArrayFacade the attributes that are used only for the pick
@@ -148,6 +150,10 @@ define([
         this._compiledShaderRotation = false;
         this._compiledShaderRotationPick = false;
 
+        this._shaderScaleByDistance = false;
+        this._compiledShaderScaleByDistance = false;
+        this._compiledShaderScaleByDistancePick = false;
+
         this._propertiesChanged = new Uint32Array(NUMBER_OF_PROPERTIES);
 
         this._maxSize = 0.0;
@@ -204,7 +210,8 @@ define([
                               BufferUsage.STATIC_DRAW, // IMAGE_INDEX_INDEX
                               BufferUsage.STATIC_DRAW, // COLOR_INDEX
                               BufferUsage.STATIC_DRAW, // ROTATION_INDEX
-                              BufferUsage.STATIC_DRAW  // ALIGNED_AXIS_INDEX
+                              BufferUsage.STATIC_DRAW, // ALIGNED_AXIS_INDEX
+                              BufferUsage.STATIC_DRAW  // SCALE_BY_DISTANCE_INDEX
                           ];
 
         var that = this;
@@ -246,6 +253,7 @@ define([
      *   horizontalOrigin : HorizontalOrigin.CENTER,
      *   verticalOrigin : VerticalOrigin.CENTER,
      *   scale : 1.0,
+     *   scaleByDistance : [5e6, 1.0, 2e7, 0.0]
      *   imageIndex : 0,
      *   color : Color.WHITE
      * });
@@ -670,6 +678,11 @@ define([
             componentsPerAttribute : 4,
             componentDatatype : ComponentDatatype.FLOAT,
             usage : buffersUsage[ROTATION_INDEX] // buffersUsage[ALIGNED_AXIS_INDEX] ignored
+        }, {
+            index : attributeIndices.scaleByDistance,
+            componentsPerAttribute : 4,
+            componentDatatype : ComponentDatatype.FLOAT,
+            usage : buffersUsage[SCALE_BY_DISTANCE_INDEX]
         }], 4 * numberOfBillboards); // 4 vertices per billboard
     }
 
@@ -849,6 +862,29 @@ define([
         writer(i + 3, rotation, x, y, z);
     }
 
+    function writeScaleByDistance(billboardCollection, context, textureAtlasCoordinates, vafWriters, billboard) {
+
+        var i = billboard._index * 4;
+        var scaleByViewerDistance = billboard.getScaleByDistance();
+
+        // if both scales are 1.0 and 1.0, then scaling by distance computation is not necessary
+        if (scaleByViewerDistance[1] !== 1.0 || scaleByViewerDistance[3] !== 1.0) {
+            billboardCollection._shaderScaleByDistance = true;
+        }
+
+        var minDistance = scaleByViewerDistance[0];
+        var scaleAtMinDist = scaleByViewerDistance[1];
+        var maxDistance = scaleByViewerDistance[2];
+        var scaleAtMaxDist = scaleByViewerDistance[3];
+
+        var allPurposeWriters = vafWriters[allPassPurpose];
+        var writer = allPurposeWriters[attributeIndices.scaleByDistance];
+        writer(i + 0, minDistance, scaleAtMinDist, maxDistance, scaleAtMaxDist);
+        writer(i + 1, minDistance, scaleAtMinDist, maxDistance, scaleAtMaxDist);
+        writer(i + 2, minDistance, scaleAtMinDist, maxDistance, scaleAtMaxDist);
+        writer(i + 3, minDistance, scaleAtMinDist, maxDistance, scaleAtMaxDist);
+    }
+
     function writeBillboard(billboardCollection, context, textureAtlasCoordinates, vafWriters, billboard) {
         writePosition(billboardCollection, context, textureAtlasCoordinates, vafWriters, billboard);
         writePixelOffset(billboardCollection, context, textureAtlasCoordinates, vafWriters, billboard);
@@ -858,6 +894,7 @@ define([
         writeOriginAndShow(billboardCollection, context, textureAtlasCoordinates, vafWriters, billboard);
         writeTextureCoordinatesAndImageSize(billboardCollection, context, textureAtlasCoordinates, vafWriters, billboard);
         writeRotationAndAlignedAxis(billboardCollection, context, textureAtlasCoordinates, vafWriters, billboard);
+        writeScaleByDistance(billboardCollection, context, textureAtlasCoordinates, vafWriters, billboard);
     }
 
     function recomputeActualPositions(billboardCollection, billboards, length, frameState, modelMatrix, recomputeBoundingVolume) {
@@ -1043,6 +1080,10 @@ define([
                     writers.push(writeRotationAndAlignedAxis);
                 }
 
+                if (properties[SCALE_BY_DISTANCE_INDEX]) {
+                    writers.push(writeScaleByDistance);
+                }
+
                 vafWriters = this._vaf.writers;
 
                 if ((billboardsToUpdateLength / billboardsLength) > 0.1) {
@@ -1117,13 +1158,17 @@ define([
                 });
             }
 
-            if (!defined(this._sp) || (this._shaderRotation && !this._compiledShaderRotation)) {
+            if (!defined(this._sp) || (this._shaderRotation && !this._compiledShaderRotation) ||
+                    (this._shaderScaleByDistance && !this._compiledShaderScaleByDistance)) {
                 this._sp = context.getShaderCache().replaceShaderProgram(
                         this._sp,
-                        (this._shaderRotation ? '#define ROTATION 1\n' : '') + BillboardCollectionVS,
+                        (this._shaderRotation ? '#define ROTATION 1\n' : '') +
+                        (this._shaderScaleByDistance ? '#define EYE_DISTANCE_SCALING 1\n' : '') +
+                        BillboardCollectionVS,
                         BillboardCollectionFS,
                         attributeIndices);
                 this._compiledShaderRotation = this._shaderRotation;
+                this._compiledShaderScaleByDistance = this._shaderScaleByDistance;
             }
 
             va = this._vaf.vaByPurpose[colorPassPurpose];
@@ -1151,13 +1196,18 @@ define([
             var pickList = this._pickCommands;
             commandLists.pickList = pickList;
 
-            if (!defined(this._spPick) || (this._shaderRotation && !this._compiledShaderRotationPick)) {
+            if (!defined(this._spPick) ||
+                    (this._shaderRotation && !this._compiledShaderRotationPick) ||
+                    (this._shaderScaleByDistance && !this._compiledShaderScaleByDistancePick)) {
                 this._spPick = context.getShaderCache().replaceShaderProgram(
                         this._spPick,
-                        (this._shaderRotation ? '#define ROTATION 1\n' : '') + '#define RENDER_FOR_PICK 1\n' + BillboardCollectionVS,
+                        (this._shaderRotation ? '#define ROTATION 1\n' : '') +
+                        (this._shaderScaleByDistance ? '#define EYE_DISTANCE_SCALING 1\n' : '') +
+                        '#define RENDER_FOR_PICK 1\n' + BillboardCollectionVS,
                         '#define RENDER_FOR_PICK 1\n' + BillboardCollectionFS,
                         attributeIndices);
-                this._compiledShaderRotation = this._shaderRotationPick;
+                this._compiledShaderRotationPick = this._shaderRotation;
+                this._compiledShaderScaleByDistancePick = this._shaderScaleByDistance;
             }
 
             va = this._vaf.vaByPurpose[pickPassPurpose];